migrated remaining tests to adoc
I migrated the remaining docs to adoc. I copied the attack potential table to all the tests (without changes except in tests 5 becasue those are largely blank so I put in "?"). I moved the remaining files to make it follow the other toolbox updates.

I moved the inventory document to adoc as well.