Clarification about AVA_VAN
Based on comments from Mary Baisch, this update attempts to clarify that that toolbox is not intended for AVA_VAN but can be used as the basis for those tests with modifications that the evaluator may decide to make.

Also found a misspelling, and added the :icons: font in the header (though I removed the section that specifically needed it).