Toolbox Template Example
This is in response to #18 about related to having a standardized template for a toolbox.

There are 4 files in the main folder (here BIO is used to mark the modality, so this should be EYE, FACE, FINGER, etc):

PAD Testing - the description of anything specific to this modality that needs to be documented (in addition to the toolbox overview) along with an introduction to the toolbox.

List - a table list of all the tests and any applicability notes (like these are only relevant with certain types of sensors or other considerations)

Inventory - a list of the inventory that is used in the test, things like paper, printers, camera, the things used to make the PAI

References - external references for any of the attacks (a master list)

Within the main folder there would then be subfoldlers marked with XX_<attack category>_attacks. The XX would be numbering for increasing difficulty (i.e. the lowest level PAI, simplest to create would be 01, the most difficult test would be the top number). The <attack category> would be some sort of title that would provide some clarity as to what the tests will be used as a source for the PAI.

Within each folder then, you would have the files names XX_YY_attack where the XX matches the folder number and YY is the test number.

The List table should match out all the numbers. While I don't have it in here, there is a (Vx) listing for some of the tests, this is when there is a number of variants for the specific test available (so a test with a V1 and V2 would have 2 variants in addition to the "base" test). This could be handled in another way though.

This used the eye tests to fill out the template, but I didn't modify any of the files.